+++ src/EntityGraphQL/Compiler/Util/ArgumentUtil.cs
@@ -28,7 +28,10 @@ public static class ArgumentUtil
                         val = Expression.Lambda((Expression)args[argField.Name], docParam).Compile().DynamicInvoke(new[] { docVariables });
                     else
                         val = args[argField.Name];
-                    propVals.Add((PropertyInfo)argField.MemberInfo!, ExpressionUtil.ChangeType(val, ((PropertyInfo)argField.MemberInfo!).PropertyType, schema));
+                    if (argField.MemberInfo is PropertyInfo info)
+                        propVals.Add((PropertyInfo)argField.MemberInfo!, ExpressionUtil.ChangeType(val, ((PropertyInfo)argField.MemberInfo!).PropertyType, schema));
+                    else
+                        fieldVals.Add((FieldInfo)argField.MemberInfo!, ExpressionUtil.ChangeType(val, ((FieldInfo)argField.MemberInfo!).FieldType, schema));
                 }
                 else
                 {