+++ src/EntityGraphQL/Schema/QueryCache.cs
@@ -5,7 +5,7 @@ using EntityGraphQL.Compiler;
 
 namespace EntityGraphQL.Schema
 {
-    internal class QueryCache
+    public class QueryCache
     {
         private readonly MemoryCache cache;
         public QueryCache()
@@ -13,24 +13,45 @@ namespace EntityGraphQL.Schema
             cache = new MemoryCache("EntityGraphQL.QueryCache");
         }
 
-        internal (GraphQLDocument?, string) GetCompiledQuery(string query)
+        public (GraphQLDocument?, string) GetCompiledQuery(string query, string? hash)
         {
-            var hash = ComputeHash(query);
-            var cached = (GraphQLDocument?)cache.Get(hash);
+            hash ??= ComputeHash(query);
+            var cached = GetCompiledQueryWithHash(hash);
             return (cached, hash);
         }
 
-        internal void AddCompiledQuery(string hash, GraphQLDocument compiledQuery)
+        public GraphQLDocument? GetCompiledQueryWithHash(string hash)
+        {
+            var cached = (GraphQLDocument?)cache.Get(hash);
+            return cached;
+        }
+
+        public void AddCompiledQuery(string hash, GraphQLDocument compiledQuery)
         {
             cache.Add(hash, compiledQuery, new CacheItemPolicy { SlidingExpiration = new System.TimeSpan(0, 10, 0) });
         }
 
-        internal string ComputeHash(string data)
+        public string ComputeHash(string data)
         {
             using SHA256 sha256Hash = SHA256.Create();
             // ComputeHash - returns byte array  
             byte[] bytes = sha256Hash.ComputeHash(Encoding.UTF8.GetBytes(data));
-            return Encoding.UTF8.GetString(bytes);
+            return ByteToHexBitFiddle(bytes);
+        }
+
+        // thanks https://stackoverflow.com/questions/311165/how-do-you-convert-a-byte-array-to-a-hexadecimal-string-and-vice-versa/14333437#14333437
+        private static string ByteToHexBitFiddle(byte[] bytes)
+        {
+            char[] c = new char[bytes.Length * 2];
+            int b;
+            for (int i = 0; i < bytes.Length; i++)
+            {
+                b = bytes[i] >> 4;
+                c[i * 2] = (char)(55 + b + (((b - 10) >> 31) & -7));
+                b = bytes[i] & 0xF;
+                c[i * 2 + 1] = (char)(55 + b + (((b - 10) >> 31) & -7));
+            }
+            return new string(c);
         }
     }
 }